+++ crates/stdto_core/src/lib.rs
@@ -9,6 +9,10 @@ pub extern crate digest;
 pub extern crate serde;
 #[cfg(feature = "json")]
 pub extern crate serde_json;
+#[cfg(feature = "yaml")]
+pub extern crate serde_yaml;
+#[cfg(feature = "toml")]
+pub extern crate toml as serde_toml;
 
 #[cfg(feature = "bytes")]
 pub use crate::{
@@ -20,7 +24,13 @@ pub use crate::{
 pub use crate::traits::ToHash;
 
 #[cfg(feature = "json")]
-pub use crate::traits::ToJson;
+pub use crate::{serde_json::Value as JsonValue, traits::ToJson};
+
+#[cfg(feature = "yaml")]
+pub use crate::{serde_yaml::Value as YamlValue, traits::ToYaml};
+
+#[cfg(feature = "toml")]
+pub use crate::{serde_toml::Value as TomlValue, traits::ToToml};
 
 #[cfg(feature = "hex")]
 pub use crate::traits::ToHex;