+++ crates/stdto_core/src/traits.rs
@@ -16,7 +16,13 @@ use bincode::Options;
 use digest::{Digest, Output};
 
 #[cfg(feature = "json")]
-use serde_json::Value;
+use serde_json::Value as JsonValue;
+
+#[cfg(feature = "yaml")]
+use serde_yaml::Value as YamlValue;
+
+#[cfg(feature = "toml")]
+use serde_toml::Value as TomlValue;
 
 macro_rules! serialize {
     (data: $self:expr, option: $option:expr) => {
@@ -362,67 +368,38 @@ pub trait ToHash: ToBytes {
 }
 
 #[cfg(feature = "json")]
-/// # A trait that can be converted to a hex string.
+/// # A trait that can de/encode to/from json.
 pub trait ToJson {
-    // --- Value ----
+    // --- JsonValue ----
     #[inline]
-    fn try_to_json_value(&self) -> Result<Value>
+    fn try_to_json_value(&self) -> Result<JsonValue>
     where
         Self: Serialize,
     {
         serde_json::to_value(self).map_err(Error::Json)
     }
     #[inline]
-    fn try_to_json_value_into(&self, writer: impl io::Write) -> Result<()>
-    where
-        Self: Serialize,
-    {
-        serde_json::to_writer(writer, self).map_err(Error::Json)
-    }
-    #[inline]
-    fn try_from_json_value(json: impl AsRef<[u8]>) -> Result<Value>
+    fn try_from_json_value(value: JsonValue) -> Result<Self>
     where
         Self: DeserializeOwned,
     {
-        serde_json::from_slice(json.as_ref()).map_err(Error::Json)
-    }
-    #[inline]
-    fn try_from_json_value_from(reader: impl io::Read) -> Result<Value>
-    where
-        Self: DeserializeOwned,
-    {
-        serde_json::from_reader(reader).map_err(Error::Json)
+        serde_json::from_value(value).map_err(Error::Json)
     }
     // --------------
     #[inline]
-    fn to_json_value(&self) -> Value
+    fn to_json_value(&self) -> JsonValue
     where
         Self: Serialize,
     {
         self.try_to_json_value().unwrap()
     }
     #[inline]
-    fn to_json_value_into(&self, writer: impl io::Write)
-    where
-        Self: Serialize,
-    {
-        self.try_to_json_value_into(writer).unwrap()
-    }
-    #[inline]
-    fn from_json_value(json: impl AsRef<[u8]>) -> Value
+    fn from_json_value(value: JsonValue) -> Self
     where
         Self: DeserializeOwned,
     {
-        Self::try_from_json_value(json).unwrap()
+        Self::try_from_json_value(value).unwrap()
     }
-    #[inline]
-    fn from_json_value_from(reader: impl io::Read) -> Value
-    where
-        Self: DeserializeOwned,
-    {
-        Self::try_from_json_value_from(reader).unwrap()
-    }
-    // --------------
 
     // --- String ----
     #[inline]
@@ -440,11 +417,11 @@ pub trait ToJson {
         serde_json::to_writer(writer, self).map_err(Error::Json)
     }
     #[inline]
-    fn try_from_json(json: impl AsRef<[u8]>) -> Result<Self>
+    fn try_from_json(json: impl AsBytes) -> Result<Self>
     where
         Self: DeserializeOwned,
     {
-        serde_json::from_slice(json.as_ref()).map_err(Error::Json)
+        serde_json::from_slice(json.as_byte_slice()).map_err(Error::Json)
     }
     #[inline]
     fn try_from_json_from(reader: impl io::Read) -> Result<Self>
@@ -469,7 +446,7 @@ pub trait ToJson {
         self.try_to_json_into(writer).unwrap()
     }
     #[inline]
-    fn from_json(json: impl AsRef<[u8]>) -> Self
+    fn from_json(json: impl AsBytes) -> Self
     where
         Self: DeserializeOwned,
     {
@@ -512,40 +489,241 @@ pub trait ToJson {
     {
         self.try_to_json_pretty_into(writer).unwrap()
     }
-    // --------------
 }
 
-#[cfg(feature = "bytes")]
-/// # A trait that can convert to a slice of bytes.
-pub trait AsBytes {
-    fn as_byte_slice(&self) -> &[u8];
+#[cfg(feature = "yaml")]
+/// # A trait that can de/encode to/from yaml.
+pub trait ToYaml {
+    // --- YamlValue ----
     #[inline]
-    fn to_bytes(&self) -> Vec<u8> {
-        self.as_byte_slice().to_vec()
-    }
-    #[inline]
-    fn into_bytes(self) -> Vec<u8>
+    fn try_to_yaml_value(&self) -> Result<YamlValue>
     where
-        Self: Sized,
+        Self: Serialize,
     {
-        self.to_bytes()
+        serde_yaml::to_value(self).map_err(Error::Yaml)
     }
     #[inline]
-    fn try_to_bytes_into(&self, mut writer: impl io::Write) -> Result<()> {
-        writer.write_all(self.as_byte_slice()).map_err(Error::Io)
+    fn try_from_yaml_value(value: YamlValue) -> Result<Self>
+    where
+        Self: DeserializeOwned,
+    {
+        serde_yaml::from_value(value).map_err(Error::Yaml)
+    }
+    // --------------
+    #[inline]
+    fn to_yaml_value(&self) -> YamlValue
+    where
+        Self: Serialize,
+    {
+        self.try_to_yaml_value().unwrap()
     }
     #[inline]
-    fn to_bytes_into(&self, writer: impl io::Write) {
-        self.try_to_bytes_into(writer).unwrap()
+    fn from_yaml_value(value: YamlValue) -> Self
+    where
+        Self: DeserializeOwned,
+    {
+        Self::try_from_yaml_value(value).unwrap()
+    }
+
+    // --- String ----
+    #[inline]
+    fn try_to_yaml(&self) -> Result<String>
+    where
+        Self: Serialize,
+    {
+        serde_yaml::to_string(self).map_err(Error::Yaml)
+    }
+    #[inline]
+    fn try_to_yaml_into(&self, writer: impl io::Write) -> Result<()>
+    where
+        Self: Serialize,
+    {
+        serde_yaml::to_writer(writer, self).map_err(Error::Yaml)
+    }
+    #[inline]
+    fn try_from_yaml(yaml: impl AsBytes) -> Result<Self>
+    where
+        Self: DeserializeOwned,
+    {
+        serde_yaml::from_slice(yaml.as_byte_slice()).map_err(Error::Yaml)
+    }
+    #[inline]
+    fn try_from_yaml_from(reader: impl io::Read) -> Result<Self>
+    where
+        Self: DeserializeOwned,
+    {
+        serde_yaml::from_reader(reader).map_err(Error::Yaml)
+    }
+    // --------------
+    #[inline]
+    fn to_yaml(&self) -> String
+    where
+        Self: Serialize,
+    {
+        self.try_to_yaml().unwrap()
+    }
+    #[inline]
+    fn to_yaml_into(&self, writer: impl io::Write)
+    where
+        Self: Serialize,
+    {
+        self.try_to_yaml_into(writer).unwrap()
+    }
+    #[inline]
+    fn from_yaml(yaml: impl AsBytes) -> Self
+    where
+        Self: DeserializeOwned,
+    {
+        Self::try_from_yaml(yaml).unwrap()
+    }
+    #[inline]
+    fn from_yaml_from(reader: impl io::Read) -> Self
+    where
+        Self: DeserializeOwned,
+    {
+        Self::try_from_yaml_from(reader).unwrap()
     }
 }
 
-#[cfg(feature = "bytes")]
-/// implement `AsBytes` for `impl AsRef<[u8]>`
-impl<T: AsRef<[u8]>> AsBytes for T {
+#[cfg(feature = "toml")]
+/// # A trait that can de/encode to/from toml.
+pub trait ToToml {
+    // --- TomlValue ----
     #[inline]
-    fn as_byte_slice(&self) -> &[u8] {
-        self.as_ref()
+    fn try_to_toml_value(&self) -> Result<TomlValue>
+    where
+        Self: Serialize,
+    {
+        TomlValue::try_from(self).map_err(Error::TomlSerialize)
+    }
+    #[inline]
+    fn try_from_toml_value(value: TomlValue) -> Result<Self>
+    where
+        Self: DeserializeOwned,
+    {
+        value.try_into().map_err(Error::TomlDeserialize)
+    }
+    // --------------
+    #[inline]
+    fn to_toml_value(&self) -> TomlValue
+    where
+        Self: Serialize,
+    {
+        self.try_to_toml_value().unwrap()
+    }
+    #[inline]
+    fn from_toml_value(value: TomlValue) -> Self
+    where
+        Self: DeserializeOwned,
+    {
+        Self::try_from_toml_value(value).unwrap()
+    }
+
+    // --- String ----
+    #[inline]
+    fn try_to_toml(&self) -> Result<String>
+    where
+        Self: Serialize,
+    {
+        toml::to_string(self).map_err(Error::TomlSerialize)
+    }
+    /// # Optimized
+    ///
+    /// This function is not well optimized yet.
+    #[inline]
+    fn try_to_toml_into(&self, mut writer: impl io::Write) -> Result<()>
+    where
+        Self: Serialize,
+    {
+        writer
+            .write_all(toml::to_vec(self)?.as_slice())
+            .map_err(Error::Io)
+    }
+    #[inline]
+    fn try_from_toml(toml: impl AsBytes) -> Result<Self>
+    where
+        Self: DeserializeOwned,
+    {
+        toml::from_slice(toml.as_byte_slice()).map_err(Error::TomlDeserialize)
+    }
+    /// # Optimized
+    ///
+    /// This function is not well optimized yet.
+    #[inline]
+    fn try_from_toml_from(mut reader: impl io::Read) -> Result<Self>
+    where
+        Self: DeserializeOwned,
+    {
+        let mut buf = Vec::new();
+        reader.read_to_end(&mut buf)?;
+        toml::from_slice(&buf).map_err(Error::TomlDeserialize)
+    }
+    // --------------
+    #[inline]
+    fn to_toml(&self) -> String
+    where
+        Self: Serialize,
+    {
+        self.try_to_toml().unwrap()
+    }
+    #[inline]
+    fn to_toml_into(&self, writer: impl io::Write)
+    where
+        Self: Serialize,
+    {
+        self.try_to_toml_into(writer).unwrap()
+    }
+    #[inline]
+    fn from_toml(toml: impl AsBytes) -> Self
+    where
+        Self: DeserializeOwned,
+    {
+        Self::try_from_toml(toml).unwrap()
+    }
+    #[inline]
+    fn from_toml_from(reader: impl io::Read) -> Self
+    where
+        Self: DeserializeOwned,
+    {
+        Self::try_from_toml_from(reader).unwrap()
+    }
+    // --------------
+    #[inline]
+    fn try_to_toml_pretty(&self) -> Result<String>
+    where
+        Self: Serialize,
+    {
+        toml::to_string_pretty(self).map_err(Error::TomlSerialize)
+    }
+    /// # Optimized
+    ///
+    /// This function is not well optimized yet.
+    #[inline]
+    fn try_to_toml_pretty_into(&self, mut writer: impl io::Write) -> Result<()>
+    where
+        Self: Serialize,
+    {
+        writer
+            .write_all(toml::to_string_pretty(self)?.as_bytes())
+            .map_err(Error::Io)
+    }
+    // --------------
+    #[inline]
+    fn to_toml_pretty(&self) -> String
+    where
+        Self: Serialize,
+    {
+        self.try_to_toml_pretty().unwrap()
+    }
+    /// # Optimized
+    ///
+    /// This function is not well optimized yet.
+    #[inline]
+    fn to_toml_pretty_into(&self, writer: impl io::Write)
+    where
+        Self: Serialize,
+    {
+        self.try_to_toml_pretty_into(writer).unwrap()
     }
 }
 
@@ -741,6 +919,40 @@ pub trait ToHex: AsBytes {
 impl<T: AsBytes> ToHex for T {}
 
 #[cfg(feature = "bytes")]
+/// # A trait that can convert to a slice of bytes.
+pub trait AsBytes {
+    fn as_byte_slice(&self) -> &[u8];
+    #[inline]
+    fn to_bytes(&self) -> Vec<u8> {
+        self.as_byte_slice().to_vec()
+    }
+    #[inline]
+    fn into_bytes(self) -> Vec<u8>
+    where
+        Self: Sized,
+    {
+        self.to_bytes()
+    }
+    #[inline]
+    fn try_to_bytes_into(&self, mut writer: impl io::Write) -> Result<()> {
+        writer.write_all(self.as_byte_slice()).map_err(Error::Io)
+    }
+    #[inline]
+    fn to_bytes_into(&self, writer: impl io::Write) {
+        self.try_to_bytes_into(writer).unwrap()
+    }
+}
+
+#[cfg(feature = "bytes")]
+/// implement `AsBytes` for `impl AsRef<[u8]>`
+impl<T: AsRef<[u8]>> AsBytes for T {
+    #[inline]
+    fn as_byte_slice(&self) -> &[u8] {
+        self.as_ref()
+    }
+}
+
+#[cfg(feature = "bytes")]
 /// # A trait for converting a byte slice to a string.
 pub trait ToStringForBytes: AsBytes {
     #[inline]
@@ -805,6 +1017,8 @@ mod tests {
     impl ToBytes for Test {}
     impl ToHash for Test {}
     impl ToJson for Test {}
+    impl ToYaml for Test {}
+    impl ToToml for Test {}
 
     #[test]
     fn test_to_bytes() {
@@ -970,6 +1184,50 @@ mod tests {
         }
     }
 
+    #[test]
+    fn test_to_yaml() {
+        let test = Test {
+            a: 1,
+            b: "hello".to_owned(),
+            c: [0; 32],
+            d: vec![1, 2, 3],
+        };
+        let yaml = test.to_yaml();
+        let test2 = Test::from_yaml(yaml);
+        if test != test2 {
+            panic!("test != test2");
+        }
+    }
+
+    #[test]
+    fn test_to_toml() {
+        let test = Test {
+            a: 1,
+            b: "hello".to_owned(),
+            c: [0; 32],
+            d: vec![1, 2, 3],
+        };
+        let json = test.to_toml();
+        let test2 = Test::from_toml(json);
+        if test != test2 {
+            panic!("test != test2");
+        }
+    }
+
+    #[test]
+    fn test_to_toml_pretty() {
+        let test = Test {
+            a: 1,
+            b: "hello".to_owned(),
+            c: [0; 32],
+            d: vec![1, 2, 3],
+        };
+        let json = test.to_toml_pretty();
+        let test2 = Test::from_toml(json);
+        if test != test2 {
+            panic!("test != test2");
+        }
+    }
     #[test]
     fn test_to_hex() {
         let bytes = b"hello world";