+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [unreleased] - ReleaseDate
 
+## BREAKING CHANGES
+
+- Some of methods in the `ToJson`.
+
+## Added
+
+- Added features `yaml` and `toml` and `file`.
+- Added `ToYaml` and `stdto::yaml` conversion.
+- Added `ToToml` and `stdto::toml` conversion.
+
+Released by [@just-do-halee](https://github.com/just-do-halee).
 
 ## [0.8.0] - 2022-12-21